+++ lib/exop/validations/type_validation.ex
@@ -100,18 +100,14 @@ defmodule Exop.TypeValidation do
   defp c(_), do: throw(:error)
 
   defp check_struct_exists(struct_name) when is_atom(struct_name) do
-    if Code.ensure_compiled?(struct_name) && function_exported?(struct_name, :__struct__, 0) do
+    with {:module, _} <- Code.ensure_compiled(struct_name),
+         true <- function_exported?(struct_name, :__struct__, 0) do
       :ok
     else
-      {:error, {:unknown_struct, struct_name}}
+      _ -> {:error, {:unknown_struct, struct_name}}
     end
   end
 
-  defp check_struct_exists(%_{}) do
-    :ok
-  end
-
-  defp check_struct_exists(unknown_struct) do
-    {:error, {:unknown_struct, unknown_struct}}
-  end
+  defp check_struct_exists(%_{}), do: :ok
+  defp check_struct_exists(unknown_struct), do: {:error, {:unknown_struct, unknown_struct}}
 end