+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.4.1] - 2020.06.08
+
+- `Code.ensure_compiled/1` instead of deprecated `Code.ensure_compiled?/1`
+- `:struct` check's bug fixed (when a value to be checked is not a struct itself)
+
 ## [1.4.0] - 2020.02.28
 
 **breaking changes in this version!**